requirements.txt
google-genai>=1.30.0
openai>=2.15.0
pillow>=10.0.0
pypdfium2>=4.30.0
python-dotenv>=1.0.1
